backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7  # 파일 업로드 처리

# Environment
python-dotenv==1.0.1

# Google Gemini
google-generativeai>=0.3.0

# Vector Store
chromadb==0.4.22
sentence-transformers==2.5.1

# Utils
pydantic==2.6.1
numpy==1.24.3
typing-extensions>=4.5.0

# Testing
pytest==8.0.2
httpx==0.26.0  # for TestClient 

# RAG
langchain>=0.1.0
langchain-community>=0.0.24
langchain-google-genai>=0.0.4
pypdf>=3.9.0  # PDF 처리
fastapi-utils>=0.2.1  # 세션 관리 